requirements.txt
datasets==1.18.3
krippendorff==0.5.0
jsonlines==2.0.0
numpy==1.21.3
pandas==1.3.4
openpyxl==3.0.9
seaborn==0.11.2
scikit-learn==1.0
spacy==3.1.3
torch==1.11.0
torchmetrics==0.7.2
toolz==0.11.2
transformers==4.17.0
sentence_transformers
pytorch_lightning==1.5.10
tqdm==4.63.0
wikiextractor==3.0.6
wordcloud==1.8.1
pyyaml
wandb==0.12.11
